Add a base class for Intellifire entities (#65077)
