+++ homeassistant/components/intellifire/binary_sensor.py
@@ -13,10 +13,10 @@ from homeassistant.components.binary_sensor import (
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from . import IntellifireDataUpdateCoordinator
 from .const import DOMAIN
+from .entity import IntellifireEntity
 
 
 @dataclass
@@ -75,27 +75,10 @@ async def async_setup_entry(
     )
 
 
-class IntellifireBinarySensor(CoordinatorEntity, BinarySensorEntity):
-    """A semi-generic wrapper around Binary Sensor entities for IntelliFire."""
+class IntellifireBinarySensor(IntellifireEntity, BinarySensorEntity):
+    """Extends IntellifireEntity with Binary Sensor specific logic."""
 
-    # Define types
-    coordinator: IntellifireDataUpdateCoordinator
     entity_description: IntellifireBinarySensorEntityDescription
-
-    def __init__(
-        self,
-        coordinator: IntellifireDataUpdateCoordinator,
-        description: IntellifireBinarySensorEntityDescription,
-    ) -> None:
-        """Class initializer."""
-        super().__init__(coordinator=coordinator)
-        self.entity_description = description
-
-        # Set the Display name the User will see
-        self._attr_name = f"Fireplace {description.name}"
-        self._attr_unique_id = f"{description.key}_{coordinator.api.data.serial}"
-        # Configure the Device Info
-        self._attr_device_info = self.coordinator.device_info
 
     @property
     def is_on(self) -> bool:
